Refacore actions and add admin command
Closes #25
